requirements.txt
allure-pytest==2.8.5
allure-python-commons==2.8.5
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.5
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.1.3
pytest-django==3.5.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
whitenoise==4.1.3
zipp==0.6.0
